1) Fixed issues with Array.GetValue() 2) Possible fix of issue where application would hang after closing main form due to threads still processing.     2.a) GNFS.SieveRelations now checks CancellationToken more frequently.

+++ GNFSCore/PrimeSignature/BitMatrix.cs
@@ -22,7 +22,7 @@ namespace GNFSCore.PrimeSignature
 
 		public BitMatrix(IEnumerable<BigInteger> array, int primeBound)
 		{
-			long maxArraySquareRoot = (long)(array.Max().SquareRoot() + 1);
+			int maxArraySquareRoot = (int)(array.Max().SquareRoot() + 1);
 			BigInteger maxArrayValue = PrimeFactory.GetValueFromIndex(maxArraySquareRoot);
 			BigInteger maxValue = BigInteger.Min(maxArrayValue, primeBound);
